[feat] 커스텀 아이콘 타입 추가 - 커스텀 아이콘 타입 추가
