+++ src/shared/Icon/customIcons/CustomIcon.tsx
@@ -1,0 +1,7 @@
+type CustomIconProps = { size: number; src: string };
+
+export function CustomIcon({ size, src }: CustomIconProps) {
+    const srcPaths = src.split("/");
+    const fileName = srcPaths[srcPaths.length - 1];
+    return <img width={size} height={size} src={src} alt={fileName}></img>;
+}
