+++ src/widgets/components/index.ts
@@ -1,4 +1,3 @@
 export { Button } from "./ui/button";
-export { Icons } from "./ui/icons";
 export { Typo } from "./ui/typo";
 export { PushNotification } from "./ui/pushNotification";
